Use single quotes in error messages to avoid issue with external tools"
Double tick quick quotes cause problems for lang_tester because we end up
wanting to write this test:

  "
  VM:
    status: error
    stdout:
      1

      ERROR: Tried loading "unknown" as a class, but failed.
  "

However that leads to a syntax error (at least in JavaSOM and yksom) since this
is two comments separated by an identifier "unknown". This seems hard to fix
without changing SOM's idea of a comment, so as a hack I think it's easier to
change the output that SOM produces.